% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exoplantname.R
\name{exoplanetname}
\alias{exoplanetname}
\title{Search for exoplanet infomation}
\usage{
exoplanetname(exoplanet)
}
\arguments{
\item{exoplanet}{a string of an existing exoplanet}
}
\value{
JSON object with infomation about the queried exoplanet
}
\description{
Exoplanet Name
}
\details{
This will return infomation about the requested exoplanet,
}
\examples{
\dontrun{
library(RCurl)
library(RJSONIO)
###Return Infomation about the exoplanet 11 Com
exoplanetname("11 Com")
###Return Infomation about the exoplanet Kepler-9
exoplanetname("Kepler-9")
###Return Infomation about all exoplants in the system
exoplanetname("")
}
}
\keyword{Name}

