% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exoplantname}
\alias{exoplantname}
\title{Search for exoplant infomation}
\usage{
exoplantname(exoplant)
}
\arguments{
\item{Star}{name will return infomation about the exoplant}
}
\value{
Infomation about the requested exoplant
}
\description{
Exoplanet Name
}
\details{
This will return infomation about the requested exoplant,
}
\examples{
###Return Infomation about the exoplanet 11 Com
exoplantname("11 Com")
###Return Infomation about the exoplanet Kepler-9
exoplantname("Kepler-9")
###Return Infomation about all exoplants in the system
exoplanetname("")
}
\keyword{Name}

