% Generated by roxygen2 (4.0.2): do not edit by hand
\name{localgroupname}
\alias{localgroupname}
\title{Search for local group galaxy infomation}
\usage{
localgroupname(localgroup)
}
\arguments{
\item{Star}{name will return infomation about the local group galaxy}
}
\value{
Infomation about the requested local group galaxy
}
\description{
Local Group Galaxy Name
}
\details{
This will return infomation about the requested local group galaxy
}
\examples{
###Return Infomation about the local group galaxy
localgroupname("IC 10")
###Return Infomation about the local group galaxy
localgroupname("WLM")
###Return Infomation about all local group galaxies in the system
localgroupname("")
}
\keyword{Name}

