% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStars-package.R
\docType{package}
\name{RStars}
\alias{RStars}
\alias{RStars-package}
\title{RStars: Access to the American Museum of Natural History's Digital Universe Dataset API}
\description{
Access to the American Museum of Natural History's Digital Universe Dataset
API, including information on positions, luminosity, color, and other data on
over 100,000 stars as well as constellations, exo-planets, clusters and 
others. Further information about this API and this dataset can be found here
\url{http://star-api.herokuapp.com/}.
}

