% Generated by roxygen2 (4.0.2): do not edit by hand
\name{starname}
\alias{starname}
\title{Search for star infomation}
\usage{
starname(star)
}
\arguments{
\item{Star}{name will return infomation about the star.}
}
\value{
Infomation about the requested star
}
\description{
Star Name
}
\details{
This will return infomation about the requested star,
}
\examples{
###Return Infomation about the Sun
starname("Sun")
###Return Infomation about the star HIP1 HD224700 Gli
starname("HIP1 HD224700 Gli")
###Return Infomation about all of the stars in the system
starname("")
}
\keyword{Name}

