% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starquery.R
\name{starquery}
\alias{starquery}
\title{Query the database for stars with specific attributes}
\usage{
starquery(queries, max_results = 1000)
}
\arguments{
\item{queries}{A \code{list} of queries to send. Refer to
\url{http://hacktheuniverse.github.io/star-api/} and example below.}

\item{max_results}{Maximum number of results to obtain (default 1000)}
}
\value{
A \code{data.frame} of query results
}
\description{
Sends a paginated query to the database and compiles the results back into a
\code{data.frame}
}
\examples{
\dontrun{
## find the (naked eye) visible stars (max 1750)
queries <- list("min[appmag]" = -2.5, "max[appmag]" = 6.5)
sq <- starquery(queries, max_results = 1750)
}
}

