% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/exoplantname.R
\name{exoplantname}
\alias{exoplantname}
\title{Search for exoplant infomation}
\usage{
exoplantname(exoplant)
}
\arguments{
\item{exoplant}{a string of an existing exoplant}
}
\value{
JSON object with infomation about the queried exoplant
}
\description{
Exoplanet Name
}
\details{
This will return infomation about the requested exoplant,
}
\examples{
\dontrun{
library(RCurl)
library(RJSONIO)
###Return Infomation about the exoplanet 11 Com
exoplantname("11 Com")
###Return Infomation about the exoplanet Kepler-9
exoplantname("Kepler-9")
###Return Infomation about all exoplants in the system
exoplantname("")
}
}
\keyword{Name}

