% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localgroupname.R
\name{localgroupname}
\alias{localgroupname}
\title{Search for local group galaxy infomation}
\usage{
localgroupname(localgroup)
}
\arguments{
\item{localgroup}{a string of an existing local group galaxy}
}
\value{
JSON object with infomation about the queried local group galaxy
}
\description{
Local Group Galaxy Name
}
\details{
This will return infomation about the requested local group galaxy
}
\examples{
\dontrun{
library(RCurl)
library(jsonlite)
###Return Infomation about the local group galaxy 
localgroupname("IC 10")
###Return Infomation about the local group galaxy 
localgroupname("WLM")
###Return Infomation about all local group galaxies in the system
localgroupname("")
}
}
\keyword{Name}

