% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/starname.R
\name{starname}
\alias{starname}
\title{Search for star infomation}
\usage{
starname(star)
}
\arguments{
\item{star}{a string of an existing star}
}
\value{
JSON object with infomation about the queried star
}
\description{
Star Name
}
\details{
This will return infomation about the requested star,
}
\examples{
\dontrun{
library(RCurl)
library(RJSONIO)
###Return Infomation about the Sun
starname("Sun")
###Return Infomation about the star HIP1 HD224700 Gli
starname("HIP1 HD224700 Gli")
###Return Infomation about all of the stars in the system
starname("")
}
}
\keyword{Name}

